# openapi.yaml
openapi: 3.0.1
info:
  title: The OpenShelf API
  version: "2.0"
  description: |
    # Summary
    The OpenShelf  API
servers:
  - url: https://TheOpenShelf.dev
    description: The OpenShelf API
paths:
...
  /hub/communities:
    get:
      operationId: getCommunities
      summary: Retrieve all communities
      tags:
        - communities-hub-api

      parameters:
        - name: searchText
          in: query
          description: Filter communities by search text in name or description
          required: false
          schema:
            type: string
        - name: location
          in: query
          description: Filter communities by location
          required: false
          schema:
            $ref: "#/components/schemas/Location"
        - name: distance
          in: query
          description: Filter communities by distance
          required: false
          schema:
            type: number
        - name: requiresApproval
          in: query
          description: Filter communities by requires approval
          required: false
          schema:
            type: boolean
        - name: isMember
          in: query
          description: Filter communities where the current user is a member
          required: false
          schema:
            type: boolean
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: A paginated list of communities
          content:
            application/json:
              schema:
                type: object
                required:
                  - totalPages
                  - totalItems
                  - currentPage
                  - itemsPerPage
                  - communities
                properties:
                  totalPages:
                    type: integer
                    example: 10
                  totalItems:
                    type: integer
                    example: 100
                  currentPage:
                    type: integer
                    example: 1
                  itemsPerPage:
                    type: integer
                    example: 10
                  communities:
                    type: array
                    items:
                      $ref: "#/components/schemas/Community"
    post:
      operationId: createCommunity
      summary: Create a new community
      tags:
        - communities-hub-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Community"
      responses:
        "201":
          description: Community created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Community"
  /hub/communities/{community-id}:
    parameters:
      - name: community-id
        in: path
        description: Community ID
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getCommunity
      summary: Retrieve a specific community by ID
      tags:
        - communities-hub-api
      responses:
        "200":
          description: A single community
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Community"
    put:
      operationId: updateCommunity
      summary: Update an existing community
      tags:
        - communities-hub-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Community"
      responses:
        "200":
          description: Community updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Community"
    delete:
      operationId: deleteCommunity
      summary: Delete a community
      tags:
        - communities-hub-api
      responses:
        "204":
          description: Community deleted successfully
  /hub/communities/{community-id}/members:
    parameters:
      - name: community-id
        in: path
        description: Community ID
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getCommunityMembers
      summary: Retrieve members of a specific community by ID
      tags:
        - communities-hub-api
      parameters:
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: A paginated list of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCommunityMembersResponse"
    post:
      operationId: addCommunityMember
      summary: Add a member to a specific community by ID
      tags:
        - communities-hub-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityMember"
      responses:
        "201":
          description: Member added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityMember"
  /hub/communities/{community-id}/members/{user-id}:
    parameters:
      - name: community-id
        in: path
        description: Community ID
        required: true
        schema:
          type: string
          format: uuid
      - name: user-id
        in: path
        description: User ID
        required: true
        schema:
          type: string
          format: uuid
    delete:
      operationId: deleteCommunityMember
      summary: Delete a member from a specific community by ID
      tags:
        - communities-hub-api
      responses:
        "204":
          description: Member deleted successfully
    put:
      operationId: updateCommunityMember
      summary: Update a member in a specific community by ID
      tags:
        - communities-hub-api
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityMember"
      responses:
        "200":
          description: Member updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityMember"
...
components:
...
  securitySchemes:
    BearerAuthOAuth:
      description: Bearer Token.
      type: http
      scheme: bearer
...
  schemas:
    User:
      type: object
      required:
        - id
        - login
        - username
        - firstName
        - lastName
        - roles
      properties:
        id:
          type: string
          format: uuid
        disabled:
          type: boolean
        isEmailVerified:
          type: boolean
        username:
          type: string
        avatarUrl:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        flatNumber:
          type: string
          example: A701
        roles:
          type: array
          items:
            type: string
            example: "admin"
        streetAddress:
          type: string
        city:
          type: string
        postalCode:
          type: string
        country:
          type: string
        preferredLanguage:
          type: string
          example: "en"
...
    PaginatedCommunityMembersResponse:
      type: object
      required:
        - totalPages
        - totalItems
        - currentPage
        - usersPerPage
        - users
      properties:
        totalPages:
          type: integer
          example: 10
        totalItems:
          type: integer
          example: 100
        currentPage:
          type: integer
          example: 1
        itemsPerPage:
          type: integer
          example: 10
        items:
          type: array
          items:
            $ref: "#/components/schemas/CommunityMember"
...
    Location:
      type: object
      required:
        - name
        - address
        - coordinates
      properties:
        name:
          type: string
        address:
          type: string
        coordinates:
          type: object
          properties:
            lat:
              type: number
            lng:
              type: number
    Community:
      type: object
      required:
        - id
        - name
        - members
        - location
        - requiresApproval
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        location:
          $ref: "#/components/schemas/Location"
        requiresApproval:
          type: boolean
        picture:
          type: string
        description:
          type: string
    CommunityMember:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            role:
              type: string
              enum: ["admin", "member", "requestingJoin"]
              example: "member"
...